# express-api/src/docs/swagger/bundled.yaml
openapi: 3.0.3
info:
  title: Main API Documentation
  description: Central documentation entry for the API.
  version: 1.0.0
paths:
...
  /api/cpp-quizzes/answers/q1:
    post:
      tags:
        - cpp-quiz
      summary: Create user answer q1
      operationId: createUserAnswerQ1
      security:
        - BearerAuth: []
      requestBody:
        description: User answer q1 information
        content:
          application/json:
            schema:
              type: object
              properties:
                answers:
                  type: object
                reasoning:
                  type: string
                confidenceLevel:
                  type: integer
        required: true
      responses:
        '201':
          description: User answer q1 created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  next:
                    type: string
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
  /api/cpp-quizzes/answers/q1-feedback:
    post:
      tags:
        - cpp-quiz
      summary: Create user answer q1 feedback
      operationId: createUserAnswerQ1Feedback
      security:
        - BearerAuth: []
      requestBody:
        description: User answer q1 feedback information
        content:
          application/json:
            schema:
              type: object
              properties:
                answers:
                  type: object
                reasoning:
                  type: string
                confidenceLevel:
                  type: integer
        required: true
      responses:
        '201':
          description: User answer q1 feedback created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
  /api/cpp-quizzes/answers/q1-discussion:
    post:
      tags:
        - cpp-quiz
      summary: Create user answer q1 discussion
      operationId: createUserAnswerQ1Discussion
      security:
        - BearerAuth: []
      requestBody:
        description: User answer q1 discussion information
        content:
          application/json:
            schema:
              type: object
              properties:
                peerInteraction:
                  type: object
                answers:
                  type: object
                reasoning:
                  type: string
                confidenceLevel:
                  type: integer
        required: true
      responses:
        '201':
          description: User answer q1 discussion created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
  /api/cpp-quizzes/answers/q2:
    post:
      tags:
        - cpp-quiz
      summary: Create user answer q2
      operationId: createUserAnswerQ2
      security:
        - BearerAuth: []
      requestBody:
        description: User answer q2 information
        content:
          application/json:
            schema:
              type: object
              properties:
                answers:
                  type: object
                confidenceLevel:
                  type: integer
        required: true
      responses:
        '201':
          description: User answer q2 created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
  /api/cpp-quizzes/answers/q3:
    post:
      tags:
        - cpp-quiz
      summary: Create user answer q3
      operationId: createUserAnswerQ3
      security:
        - BearerAuth: []
      requestBody:
        description: User answer q3 information
        content:
          application/json:
            schema:
              type: object
              properties:
                answers:
                  type: object
                confidenceLevel:
                  type: integer
        required: true
      responses:
        '201':
          description: User answer q3 created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
  /api/cpp-quizzes/questions:
    post:
      tags:
        - cpp-quiz
      summary: Create a question
      operationId: createQuestion
      security:
        - BearerAuth: []
      requestBody:
        description: Question information
        content:
          application/json:
            schema:
              type: object
              properties:
                correctAnswers:
                  type: object
                reasoning:
                  type: string
                week:
                  type: integer
                round:
                  type: integer
                number:
                  type: integer
        required: true
      responses:
        '201':
          description: Question created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
  /api/cpp-quizzes/questions/current:
    get:
      tags:
        - cpp-quiz
      summary: Get current question info
      operationId: getCurrentQuestionInfo
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Get current question info successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  imageUrl:
                    type: string
                  answersCount:
                    type: integer
                  submitEndpoint:
                    type: string
                  systemFeedback:
                    type: string
                  peerInteraction:
                    type: object
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...